Move spatial_hash.hpp to algorithm
