+++ beluga/include/beluga/beluga.hpp
@@ -23,7 +23,6 @@
 #include <beluga/algorithm.hpp>
 #include <beluga/motion.hpp>
 #include <beluga/sensor.hpp>
-#include <beluga/spatial_hash.hpp>
 #include <beluga/tuple_vector.hpp>
 #include <beluga/type_traits.hpp>
 #include <beluga/views.hpp>
